+++ src/pytools/api/_api.py
@@ -18,6 +18,7 @@ from typing import (
     TypeVar,
     Union,
     cast,
+    overload,
 )
 
 import numpy as np
@@ -49,7 +50,6 @@ __all__ = [
 
 T = TypeVar("T")
 T_Collection = TypeVar("T_Collection", list, set, tuple)
-T_Callable = TypeVar("T_Callable", bound=Callable)
 T_Iterable = TypeVar("T_Iterable", bound=Iterable)
 T_Type = TypeVar("T_Type", bound=type)
 
@@ -79,8 +79,8 @@ def is_list_like(obj: Any) -> bool:
     - :class:`str`
     - :class:`bytes`
     - :class:`pandas.DataFrame`: inconsistent behaviour of the sequence interface;
-        iterating a data frame yields the values of the column index, while the length
-        of a data frame is its number of rows
+      iterating a data frame yields the values of the column index, while the length
+      of a data frame is its number of rows
     - :class:`numpy.ndarray` instances with 0 dimensions
 
     :param obj: The object to check
@@ -98,6 +98,30 @@ def is_list_like(obj: Any) -> bool:
     )
 
 
+@overload
+def to_tuple(
+    values: Iterable[T],
+    *,
+    element_type: Optional[Union[Type[T], Tuple[Type[T], ...]]] = None,
+    optional: bool = False,
+    arg_name: Optional[str] = None,
+) -> Tuple[T, ...]:
+    """see below for implementation"""
+    pass
+
+
+@overload
+def to_tuple(
+    values: Optional[T],
+    *,
+    element_type: Optional[Union[Type[T], Tuple[Type[T], ...]]] = None,
+    optional: bool = False,
+    arg_name: Optional[str] = None,
+) -> Tuple[T, ...]:
+    """see below for implementation"""
+    pass
+
+
 def to_tuple(
     values: Union[Iterable[T], T, None],
     *,
@@ -106,12 +130,12 @@ def to_tuple(
     arg_name: Optional[str] = None,
 ) -> Tuple[T, ...]:
     """
-    Return the given values as a tuple.
+    Return the given values as a tuple:
 
-    - if arg `values` is a tuple, return arg `values` unchanged
-    - if arg `values` is an iterable other than a tuple, return a tuple of its elements
-    - if arg `values` is not an iterable, return a tuple with the value as its only
-      element
+    - If arg `values` is a tuple, return arg `values` unchanged.
+    - If arg `values` is an iterable other than a tuple, return a tuple of its elements.
+    - If arg `values` is not an iterable, return a tuple with the value as its only
+      element.
 
     :param values: one or more elements to return as a tuple
     :param element_type: expected type of the values, or a tuple of alternative types
@@ -135,6 +159,30 @@ def to_tuple(
     )
 
 
+@overload
+def to_list(
+    values: Iterable[T],
+    *,
+    element_type: Optional[Union[Type[T], Tuple[Type[T], ...]]] = None,
+    optional: bool = False,
+    arg_name: Optional[str] = None,
+) -> List[T]:
+    """see below for implementation"""
+    pass
+
+
+@overload
+def to_list(
+    values: Optional[T],
+    *,
+    element_type: Optional[Union[Type[T], Tuple[Type[T], ...]]] = None,
+    optional: bool = False,
+    arg_name: Optional[str] = None,
+) -> List[T]:
+    """see below for implementation"""
+    pass
+
+
 @subsdoc(pattern="tuple", replacement="list", using=to_tuple)
 def to_list(
     values: Union[Iterable[T], T, None],
@@ -155,6 +203,30 @@ def to_list(
     )
 
 
+@overload
+def to_set(
+    values: Iterable[T],
+    *,
+    element_type: Optional[Union[Type[T], Tuple[Type[T], ...]]] = None,
+    optional: bool = False,
+    arg_name: Optional[str] = None,
+) -> Set[T]:
+    """see below for implementation"""
+    pass
+
+
+@overload
+def to_set(
+    values: Optional[T],
+    *,
+    element_type: Optional[Union[Type[T], Tuple[Type[T], ...]]] = None,
+    optional: bool = False,
+    arg_name: Optional[str] = None,
+) -> Set[T]:
+    """see below for implementation"""
+    pass
+
+
 @subsdoc(pattern="tuple", replacement="set", using=to_tuple)
 def to_set(
     values: Union[Iterable[T], T, None],
@@ -173,6 +245,30 @@ def to_set(
         optional=optional,
         arg_name=arg_name,
     )
+
+
+@overload
+def to_collection(
+    values: Iterable[T],
+    *,
+    element_type: Optional[Union[Type[T], Tuple[Type[T], ...]]] = None,
+    optional: bool = False,
+    arg_name: Optional[str] = None,
+) -> Collection[T]:
+    """see below for implementation"""
+    pass
+
+
+@overload
+def to_collection(
+    values: Optional[T],
+    *,
+    element_type: Optional[Union[Type[T], Tuple[Type[T], ...]]] = None,
+    optional: bool = False,
+    arg_name: Optional[str] = None,
+) -> Collection[T]:
+    """see below for implementation"""
+    pass
 
 
 @subsdoc(pattern="iterable other than a collection", replacement="iterable")
@@ -233,7 +329,9 @@ def _to_collection(
 
     if element_type:
         validate_element_types(
-            elements, expected_type=element_type, name=f"arg {arg_name}"
+            elements,
+            expected_type=element_type,
+            name=f"arg {arg_name}" if arg_name else None,
         )
 
     return elements
@@ -265,34 +363,15 @@ def validate_type(
     if optional and value is None:
         return None
 
-    if isinstance(value, expected_type):
-        return value
-    else:
-        if name:
-            message_head = f"{name} requires"
-        else:
-            message_head = "expected"
-
-        expected_type_tuple: Tuple[type, ...]
-
-        if not isinstance(expected_type, tuple):
-            expected_type_tuple = (expected_type,)
-        else:
-            expected_type_tuple = expected_type
-        if optional:
-            expected_type_tuple = (*expected_type_tuple, type(None))
-
-        expected_type_str = " or ".join(t.__name__ for t in expected_type_tuple)
-
-        observed_type = type(value).__name__
-
-        # noinspection SpellCheckingInspection
-        det = "an" if observed_type[0] in "aeiou" else "a"
-
-        raise TypeError(
-            f"{message_head} an instance of {expected_type_str} "
-            f"but got {det} {observed_type}"
+    if not isinstance(value, expected_type):
+        _raise_type_mismatch(
+            name=name,
+            expected_type=_as_optional_type(expected_type, is_optional=optional),
+            mismatched_type=type(value),
+            is_single=True,
         )
+
+    return value
 
 
 def validate_element_types(
@@ -323,31 +402,56 @@ def validate_element_types(
         )
 
     if expected_type is not object:
-        if optional:
-            expected_type = (
-                (*expected_type, type(None))
-                if isinstance(expected_type, tuple)
-                else (expected_type, type(None))
-            )
+        expected_type = _as_optional_type(type_=expected_type, is_optional=optional)
+
         for element in iterable:
             if not isinstance(element, expected_type):
-                if name:
-                    message_head = f"{name} requires"
-                else:
-                    message_head = "expected"
-
-                if isinstance(expected_type, type):
-                    expected_type_str = expected_type.__name__
-                else:
-                    expected_type_str = (
-                        f"one of {{{', '.join(t.__name__ for t in expected_type)}}}"
-                    )
-                raise TypeError(
-                    f"{message_head} instances of {expected_type_str} "
-                    f"but got a {type(element).__name__}"
+                _raise_type_mismatch(
+                    name=name,
+                    expected_type=expected_type,
+                    mismatched_type=type(element),
+                    is_single=False,
                 )
 
     return iterable
+
+
+def _as_optional_type(
+    type_: Union[type, Tuple[type, ...]], *, is_optional: bool
+) -> Union[type, Tuple[type, ...]]:
+    # if is_optional is True, return a tuple comprising the original (types) and None
+    if is_optional:
+        if isinstance(type_, tuple):
+            return (*type_, type(None))
+        else:
+            return (type_, type(None))
+    else:
+        return type_
+
+
+def _raise_type_mismatch(
+    *,
+    name: Optional[str],
+    expected_type: Union[type, Tuple[type, ...]],
+    mismatched_type: type,
+    is_single: bool,
+) -> None:
+    if name:
+        message_head = f"{name} requires"
+    else:
+        message_head = "expected"
+
+    if isinstance(expected_type, type):
+        expected_type_str = expected_type.__name__
+    else:
+        expected_type_str = f"one of {{{', '.join(t.__name__ for t in expected_type)}}}"
+
+    instance = "an instance" if is_single else "instances"
+
+    raise TypeError(
+        f"{message_head} {instance} of {expected_type_str} "
+        f"but got: {mismatched_type.__name__}"
+    )
 
 
 def get_generic_bases(class_: type) -> Tuple[type, ...]:
@@ -373,12 +477,12 @@ def get_generic_bases(class_: type) -> Tuple[type, ...]:
 
 
 def deprecated(
-    function: Optional[T_Callable] = None, *, message: Optional[str] = None
-) -> Union[T_Callable, Callable[[T_Callable], T_Callable]]:
+    function: Optional[Callable] = None, *, message: Optional[str] = None
+) -> Callable:
     """
     Decorator to mark a function as deprecated.
 
-    Logs a warning when the decorated function is called.
+    Issues a warning when the decorated function is called.
 
     Usage:
 
@@ -403,7 +507,7 @@ def deprecated(
         if not callable(func):
             raise ValueError("Deprecated object must be callable")
 
-    def _deprecated_inner(func: T_Callable) -> T_Callable:
+    def _deprecated_inner(func: Callable) -> Callable:
         _validate_function(func)
 
         @wraps(func)
@@ -422,7 +526,7 @@ def deprecated(
                 )
             return func(*args, **kwargs)
 
-        return cast(T_Callable, new_func)
+        return new_func
 
     if function is None:
         return _deprecated_inner