+++ src/pytools/api/_doc_validator.py
@@ -74,27 +74,27 @@ class DocValidator:
     is provided in parameter ``validate_protected``.
     """
 
-    #: the root directory of all Python files to be validated
+    #: The root directory of all Python files to be validated.
     root_dir: str
 
-    #: names of protected functions and methods to be validated
+    #: Names of protected functions and methods to be validated.
     validate_protected: Tuple[str, ...]
 
-    #: names of modules for which parameter documentation and type hints should not be
-    #: validated
+    #: Names of modules for which parameter documentation and type hints should not be
+    #: validated.
     exclude_from_parameter_validation: Optional[Pattern]
 
-    #: after validation, lists all errors per definition
+    #: A dictionary mapping definitions to lists of errors identified during validation.
     validation_errors: Dict[str, List[str]]
 
-    #: tests to run on each definition during validation
+    #: The documentation tests to run on each definition during validation.
     validation_tests: Tuple[DocTest, ...]
 
-    #: default value for parameter ``validate_protected``
+    #: The default value for parameter ``validate_protected``.
     DEFAULT_VALIDATE_PROTECTED = ("__init__",)
 
-    #: default doc tests to run
-    DEFAULT_DOC_TESTS: Tuple[DocTest, DocTest, DocTest, DocTest] = (
+    #: Default collection of documentation tests to run during validation.
+    DEFAULT_VALIDATION_TESTS: Tuple[DocTest, ...] = (
         HasDocstring(),
         HasMatchingParameterDoc(),
         HasWellFormedDocstring(),
@@ -138,9 +138,9 @@ class DocValidator:
 
         self.validation_errors = {}
         self.validation_tests = (
-            (*self.DEFAULT_DOC_TESTS, *additional_tests)
+            (*self.DEFAULT_VALIDATION_TESTS, *additional_tests)
             if additional_tests
-            else self.DEFAULT_DOC_TESTS
+            else self.DEFAULT_VALIDATION_TESTS
         )
 
     __init__.__doc__ = __init__.__doc__.replace(  # type: ignore