+++ src/pytools/api/doc/_doc.py
@@ -211,6 +211,9 @@ class FunctionDefinition(NamedElementDefinition[FunctionType]):
     #: name of the special "return" parameter used in signatures and type annotations
     PARAM_RETURN = "return"
 
+    # defined in superclass, repeated here for Sphinx
+    element: FunctionType
+
     def list_actual_parameters(self, include_return: bool) -> List[str]:
         """
         Extract all parameter names from the function signature
@@ -257,18 +260,21 @@ class DocTest(metaclass=ABCMeta):
         """
         pass
 
+    def __repr__(self) -> str:
+        return type(self).__name__ + "()"
+
 
 @inheritdoc(match="""[see superclass]""")
 class HasDocstring(DocTest):
     """
-    Test that the definition's docstring is defined and not empty.
+    Test that the definition's docstring exists and is not empty.
     """
 
     def test(self, definition: APIDefinition) -> Union[None, str, List[str]]:
         """[see superclass]"""
 
         doc = definition.docstring
-        if not (doc and str(doc).strip()) and not definition.name == "__init__":
+        if not (doc and str(doc).strip()) and definition.name != "__init__":
             return "missing docstring"
         else:
             return None