+++ src/pytools/parallelization/_parallelization.py
@@ -39,7 +39,7 @@ __all__ = [
     "Job",
     "JobQueue",
     "JobRunner",
-    "NestedQueue",
+    "CompositeQueue",
     "ParallelizableMixin",
     "SimpleQueue",
 ]
@@ -68,8 +68,22 @@ __tracker = AllTracker(globals())
 
 class ParallelizableMixin:
     """
-    Mix-in class that supports parallelizing one or more operations using joblib.
+    Mix-in class that supports parallelizing one or more operations using :mod:`joblib`.
     """
+
+    #: Number of jobs to use in parallel; if ``None``, use joblib default.
+    n_jobs: Optional[int]
+
+    #: If ``True``, use threads in the parallel runs;
+    #: if ``False`` or ``None``, use multiprocessing.
+    shared_memory: Optional[bool]
+
+    #: Number of batches to pre-dispatch; if ``None``, use joblib default.
+    pre_dispatch: Optional[Union[str, int]]
+
+    #: Verbosity level used in the parallel computation;
+    #: if ``None``, use joblib default.
+    verbose: Optional[int]
 
     def __init__(
         self,
@@ -83,25 +97,16 @@ class ParallelizableMixin:
         :param n_jobs: number of jobs to use in parallel;
             if ``None``, use joblib default (default: ``None``)
         :param shared_memory: if ``True``, use threads in the parallel runs; if
-            ``False``, use multiprocessing (default: ``False``)
+            ``False`` or ``None``, use multiprocessing (default: ``None``)
         :param pre_dispatch: number of batches to pre-dispatch;
             if ``None``, use joblib default (default: ``None``)
         :param verbose: verbosity level used in the parallel computation;
             if ``None``, use joblib default (default: ``None``)
         """
         super().__init__()
-        #: Number of jobs to use in parallel; if ``None``, use joblib default.
         self.n_jobs = n_jobs
-
-        #: If ``True``, use threads in the parallel runs;
-        #: if ``False``, use multiprocessing.
         self.shared_memory = shared_memory
-
-        #: Number of batches to pre-dispatch; if ``None``, use joblib default.
         self.pre_dispatch = pre_dispatch
-
-        #: Verbosity level used in the parallel computation;
-        #: if ``None``, use joblib default.
         self.verbose = verbose
 
         self._parallel_kwargs = {
@@ -211,6 +216,18 @@ class JobRunner(ParallelizableMixin):
     """
     Runs job queues in parallel and aggregates results.
     """
+
+    # defined in superclass, repeated here for Sphinx
+    n_jobs: Optional[int]
+
+    # defined in superclass, repeated here for Sphinx
+    shared_memory: Optional[bool]
+
+    # defined in superclass, repeated here for Sphinx
+    pre_dispatch: Optional[Union[str, int]]
+
+    # defined in superclass, repeated here for Sphinx
+    verbose: Optional[int]
 
     @classmethod
     def from_parallelizable(
@@ -331,6 +348,9 @@ class SimpleQueue(
     A simple queue, running a given list of jobs.
     """
 
+    # defined in superclass, repeated here for Sphinx
+    lock: LockType
+
     #: The jobs run by this queue.
     _jobs: Tuple[Job[T_Job_Result], ...]
 
@@ -339,7 +359,9 @@ class SimpleQueue(
         :param jobs: jobs to be run by this queue in the given order
         """
         super().__init__()
-        self._jobs = to_tuple(jobs, element_type=Job, arg_name="jobs")
+        self._jobs = to_tuple(
+            jobs, element_type=cast(Type[Job[T_Job_Result]], Job), arg_name="jobs"
+        )
 
     def jobs(self) -> Iterable[Job[T_Job_Result]]:
         """[see superclass]"""
@@ -350,11 +372,13 @@ class SimpleQueue(
 
 
 @inheritdoc(match="""[see superclass]""")
-class NestedQueue(JobQueue[T_Job_Result, List[T_Job_Result]], Generic[T_Job_Result]):
+class CompositeQueue(JobQueue[T_Job_Result, List[T_Job_Result]], Generic[T_Job_Result]):
     """
-    Runs all jobs in a given list of compatible queues and returns their results as a
-    flat list.
+    A queue composed from a collection of compatible queues.
     """
+
+    # defined in superclass, repeated here for Sphinx
+    lock: LockType
 
     #: The queues run by this queue.
     queues: Tuple[JobQueue[T_Job_Result, List[T_Job_Result]], ...]
@@ -363,10 +387,17 @@ class NestedQueue(JobQueue[T_Job_Result, List[T_Job_Result]], Generic[T_Job_Resu
         self, queues: Sequence[JobQueue[T_Job_Result, List[T_Job_Result]]]
     ) -> None:
         """
-        :param queues: queues to be run by this queue in the given order
+        :param queues: queues whose elements will be added to this queue in the given
+            order
         """
         super().__init__()
-        self.queues = to_tuple(queues)
+        self.queues = to_tuple(
+            queues,
+            element_type=cast(
+                Type[JobQueue[T_Job_Result, List[T_Job_Result]]], JobQueue
+            ),
+            arg_name="queues",
+        )
 
     def jobs(self) -> Iterable[Job[T_Job_Result]]:
         """[see superclass]"""