+++ test/test/pytools/viz/test_color.py
@@ -2,6 +2,13 @@ import pytest
 
 from pytools.viz.color import RgbaColor, RgbColor
 
+MSG_INCOMPLETE_RGB_KWARG = (
+    "incomplete RGB keyword arguments: need to provide r, g, and b"
+)
+MSG_MUST_BE_A_STRING = "single color argument must be a string"
+MSG_NEED_COLOR_NAME_OR_RGB_VALUES = "need 1 color name or 3 RGB values"
+MSG_INVALID_RGBA_VALUES = "invalid RGBA values"
+
 
 def test_rgb() -> None:
     red = (1.0, 0.0, 0.0)
@@ -11,45 +18,45 @@ def test_rgb() -> None:
     assert RgbColor(1, 0, 0) == red
     assert RgbColor(r=1, g=0, b=0) == red
 
-    with pytest.raises(ValueError, match="need 1 color name or 3 RGB values"):
-        RgbColor("red", 1)
-    with pytest.raises(ValueError, match="need 1 color name or 3 RGB values"):
-        RgbColor(1, 1)
-    with pytest.raises(ValueError, match="need 1 color name or 3 RGB values"):
+    with pytest.raises(ValueError, match=MSG_NEED_COLOR_NAME_OR_RGB_VALUES):
+        RgbColor("red", 1)  # type: ignore
+    with pytest.raises(ValueError, match=MSG_NEED_COLOR_NAME_OR_RGB_VALUES):
+        RgbColor(1, 1)  # type: ignore
+    with pytest.raises(ValueError, match=MSG_NEED_COLOR_NAME_OR_RGB_VALUES):
         # noinspection PyArgumentList
-        RgbColor(1, 0, 2, 4)
+        RgbColor(1, 0, 2, 4)  # type: ignore
 
     with pytest.raises(
         ValueError,
         match="mixed use of positional and keyword arguments for color arguments",
     ):
         # noinspection PyArgumentList
-        RgbColor("red", c=1)
+        RgbColor("red", c=1)  # type: ignore
 
     with pytest.raises(ValueError, match="mixed use of named color and color channels"):
         # noinspection PyArgumentList
-        RgbColor(r=1, g=1, b=1, c=0)
+        RgbColor(r=1, g=1, b=1, c=0)  # type: ignore
 
-    with pytest.raises(ValueError, match="single color argument must be a string"):
+    with pytest.raises(ValueError, match=MSG_MUST_BE_A_STRING):
         # noinspection PyTypeChecker
-        RgbColor(1)
+        RgbColor(1)  # type: ignore
 
-    with pytest.raises(ValueError, match="invalid RGBA values"):
-        RgbColor(1, 0, 2)
-
-    with pytest.raises(
-        ValueError,
-        match="incomplete RGB keyword arguments: need to provide r, g, and b",
-    ):
-        # noinspection PyArgumentList
-        RgbColor(r=1, g=0)
+    with pytest.raises(ValueError, match=MSG_INVALID_RGBA_VALUES):
+        RgbColor(1, 0, 2)  # type: ignore
 
     with pytest.raises(
         ValueError,
-        match="incomplete RGB keyword arguments: need to provide r, g, and b",
+        match=MSG_INCOMPLETE_RGB_KWARG,
     ):
         # noinspection PyArgumentList
-        RgbColor(r=1, c=0)
+        RgbColor(r=1, g=0)  # type: ignore
+
+    with pytest.raises(
+        ValueError,
+        match=MSG_INCOMPLETE_RGB_KWARG,
+    ):
+        # noinspection PyArgumentList
+        RgbColor(r=1, c=0)  # type: ignore
 
 
 def test_rgba() -> None:
@@ -71,39 +78,39 @@ def test_rgba() -> None:
 
     with pytest.raises(ValueError, match="need 3 RGB values or 4 RGBA values"):
         # noinspection PyArgumentList
-        RgbaColor(1, 1, 1, 1, 1)
+        RgbaColor(1, 1, 1, 1, 1)  # type: ignore
 
     with pytest.raises(
         ValueError,
         match=r"all color arguments must be numeric, but are: \('red', 1, 1, 1, 1\)",
     ):
         # noinspection PyArgumentList
-        RgbaColor("red", 1, 1, 1, 1)
+        RgbaColor("red", 1, 1, 1, 1)  # type: ignore
 
     with pytest.raises(
         ValueError,
         match="mixed use of positional and keyword arguments for color arguments",
     ):
         # noinspection PyArgumentList
-        RgbaColor("red", c=1, alpha=0)
+        RgbaColor("red", c=1, alpha=0)  # type: ignore
 
     with pytest.raises(ValueError, match="mixed use of named color and color channels"):
         # noinspection PyArgumentList
-        RgbaColor(r=1, g=1, b=1, c=0, alpha=0)
+        RgbaColor(r=1, g=1, b=1, c=0, alpha=0)  # type: ignore
 
-    with pytest.raises(ValueError, match="single color argument must be a string"):
+    with pytest.raises(ValueError, match=MSG_MUST_BE_A_STRING):
         # noinspection PyTypeChecker
-        RgbaColor(1)
-    with pytest.raises(ValueError, match="single color argument must be a string"):
+        RgbaColor(1)  # type: ignore
+    with pytest.raises(ValueError, match=MSG_MUST_BE_A_STRING):
         # noinspection PyTypeChecker
-        RgbaColor(1, 1)
-    with pytest.raises(ValueError, match="single color argument must be a string"):
+        RgbaColor(1, 1)  # type: ignore
+    with pytest.raises(ValueError, match=MSG_MUST_BE_A_STRING):
         # noinspection PyTypeChecker
-        RgbaColor(1, alpha=0)
+        RgbaColor(1, alpha=0)  # type: ignore
 
-    with pytest.raises(ValueError, match="invalid RGBA values"):
+    with pytest.raises(ValueError, match=MSG_INVALID_RGBA_VALUES):
         RgbaColor(1, 0, 2)
-    with pytest.raises(ValueError, match="invalid RGBA values"):
+    with pytest.raises(ValueError, match=MSG_INVALID_RGBA_VALUES):
         RgbaColor(1, 0, 1, alpha=2)
 
     with pytest.raises(
@@ -111,4 +118,4 @@ def test_rgba() -> None:
         match="incomplete RGB keyword arguments: need to provide r, g, and b",
     ):
         # noinspection PyArgumentList
-        RgbaColor(r=1, g=0, alpha=1)
+        RgbaColor(r=1, g=0, alpha=1)  # type: ignore